+++ src/ResourceManager/Resources/ChangeLog.md
@@ -24,6 +24,7 @@
 * Fix roleassignment get calls when there are duplicate objectIds.
 * Fix RoleAssignment get to respect the ExpandPrincipalGroups parameter
 * Fix Roleassignment get calls to be used with roledefinition ID.
+* Add Support for Dataactions and not dataactions to be passed in roledefinition create calls
 
 ## Version 5.3.0
 * Get-AzureRmADServicePrincipal: Removed -ServicePrincipalName from the default Empty parameter set as it was redundant with the SPN parameter set